-create IndustryViewHolder.
