+++ app/src/main/java/ru/practicum/android/diploma/presentation/filtration/industry/IndustryAdapter.kt
@@ -2,12 +2,10 @@ package ru.practicum.android.diploma.presentation.filtration.industry
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
-import android.widget.RadioButton
-import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
 import ru.practicum.android.diploma.databinding.ItemIndustryViewBinding
 
-class IndustryAdapter : RecyclerView.Adapter<IndustryAdapter.IndustryViewHolder>() {
+class IndustryAdapter : RecyclerView.Adapter<IndustryViewHolder>() {
     private var industryList = arrayListOf<Industry>()
     private var itemIndustryClickListener: ItemIndustryClickInterface? = null
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): IndustryViewHolder {
@@ -35,37 +33,7 @@ class IndustryAdapter : RecyclerView.Adapter<IndustryAdapter.IndustryViewHolder>
         this.itemIndustryClickListener = itemClickListener
     }
 
-    class IndustryViewHolder(
-        binding: ItemIndustryViewBinding,
-        private val itemIndustryClickListener: ItemIndustryClickInterface?
-    ) : RecyclerView.ViewHolder(binding.root) {
-        private val industryName: TextView = binding.textViewIndustry
-        private val industryRadioButton: RadioButton = binding.radioButtonIndustry
-        private var itemIndustry: Industry? = null
-
-        init {
-            // меняем состояние радиокнопки
-            industryRadioButton.isChecked = !industryRadioButton.isChecked
-            itemView.setOnClickListener {
-                if (itemIndustry != null) {
-                    itemIndustryClickListener?.onItemIndustryClick(
-                        // записываем и отправляем индастри с новым состоянием радиокнопки
-                        Industry(
-                            itemIndustry!!.name,
-                            industryRadioButton.isChecked
-                        )
-                    )
-                }
-            }
-        }
-
-        fun bind(industry: Industry) {
-            itemIndustry = industry
-            industryName.text = industry.name
-            industryRadioButton.isChecked = industry.check
-        }
-    }
-
+    // удалить этот временный data class Industry когда будет готов нормальный
     data class Industry(
         val name: String,
         val check: Boolean
